requirements.txt
Django>=5.2.6
requests>=2.31.0
Pillow>=10.0.0
zipstream-ng>=1.6.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-widget-tweaks>=1.5.0
python-slugify>=8.0.1
pandas>=2.0.0
environs>=10.0.0
